+++ site/src/other.rs
@@ -109,6 +109,25 @@ pub fn Technical() -> impl IntoView {
 }
 
 #[component]
+pub fn StackIdioms() -> impl IntoView {
+    view! {
+        <h1>"Common Stack Idioms"</h1>
+        <p>"This page contains some common stack idioms that you may find useful."</p>
+        <p>"They are presented as rearrangements of numbers which are then grouped into an array so that you can see the result."</p>
+        <Editor example="[. 1]"/>
+        <Editor example="[∶ 1 2]"/>
+        <Editor example="[, 1 2]"/>
+        <Editor example="[; 1 2]"/>
+        <Editor example="[,, 1 2]"/>
+        <Editor example="[,∶ 1 2]"/>
+        <Editor example="[⊙. 1 2]"/>
+        <Editor example="[⊙; 1 2]"/>
+        <Editor example="[⊙∶ 1 2 3]"/>
+        <Editor example="[⊙, 1 2 3]"/>
+    }
+}
+
+#[component]
 pub fn Install() -> impl IntoView {
     view! {
         <h2>"Installing Uiua"</h2>